Remove more special casing of content URIs
